cast str with rev_map
