+++ crates/polars-core/src/chunked_array/logical/categorical/mod.rs
@@ -2,6 +2,7 @@ mod builder;
 mod from;
 mod merge;
 mod ops;
+pub mod revmap;
 pub mod string_cache;
 
 use bitflags::bitflags;
@@ -9,6 +10,7 @@ pub use builder::*;
 pub use merge::*;
 use polars_utils::iter::EnumerateIdxTrait;
 use polars_utils::sync::SyncPtr;
+pub use revmap::*;
 
 use super::*;
 use crate::chunked_array::Settings;
@@ -109,16 +111,17 @@ impl CategoricalChunked {
             RevMapping::Local(categories, _) => categories,
             RevMapping::Enum(categories, _) => categories,
         };
-        let physical = self.physical();
 
-        let mut builder =
-            CategoricalChunkedBuilder::new(self.name(), physical.len(), self.get_ordering());
-        let iter = physical
-            .downcast_iter()
-            .map(|z| z.into_iter().map(|z| z.copied()))
-            .collect::<Vec<_>>();
-        builder.global_map_from_local(iter, self.len(), categories.clone());
-        Ok(builder.finish())
+        // Safety: keys and values are in bounds
+        unsafe {
+            Ok(CategoricalChunked::from_keys_and_values_global(
+                self.name(),
+                self.physical.into_iter(),
+                self.len(),
+                categories,
+                self.get_ordering(),
+            ))
+        }
     }
 
     // Convert to fixed enum. In case a value is not in the categories return Error
@@ -186,25 +189,6 @@ impl CategoricalChunked {
     /// Set flags for the Chunked Array
     pub(crate) fn set_flags(&mut self, flags: Settings) {
         self.physical_mut().set_flags(flags)
-    }
-
-    /// Build a categorical from an original RevMap. That means that the number of categories in the `RevMapping == self.unique().len()`.
-    pub(crate) fn from_chunks_original(
-        name: &str,
-        chunk: PrimitiveArray<u32>,
-        rev_map: RevMapping,
-        ordering: CategoricalOrdering,
-    ) -> Self {
-        let ca = ChunkedArray::with_chunk(name, chunk);
-        let mut logical = Logical::<UInt32Type, _>::new_logical::<CategoricalType>(ca);
-        logical.2 = Some(DataType::Categorical(Some(Arc::new(rev_map)), ordering));
-
-        let mut bit_settings = BitSettings::default();
-        bit_settings.insert(BitSettings::ORIGINAL);
-        Self {
-            physical: logical,
-            bit_settings,
-        }
     }
 
     /// Return whether or not the [`CategoricalChunked`] uses the lexical order
