+++ crates/polars-ops/src/series/ops/cut.rs
@@ -49,11 +49,13 @@ fn map_cats(
         let outvals = vec![brk_vals.finish().into_series(), bld.finish().into_series()];
         Ok(StructChunked::new(&out_name, &outvals)?.into_series())
     } else {
-        bld.drain_iter(s_iter.map(|opt| {
-            opt.filter(|x| !x.is_nan())
-                .map(|x| unsafe { *cl.get_unchecked(sorted_breaks.partition_point(|v| op(&x, v))) })
-        }));
-        Ok(bld.finish().into_series())
+        Ok(bld
+            .drain_iter_and_finish(s_iter.map(|opt| {
+                opt.filter(|x| !x.is_nan()).map(|x| unsafe {
+                    *cl.get_unchecked(sorted_breaks.partition_point(|v| op(&x, v)))
+                })
+            }))
+            .into_series())
     }
 }
